Organize the code structure and add debug code
