+++ blockchain/txvalidator.go
@@ -767,7 +767,7 @@ func (b *BlockChain) checkTransactionOutput(txn *Transaction,
 	if b.GetHeight() >= b.chainParams.PublicDPOSHeight && specialOutputCount > 1 {
 		return errors.New("special output count should less equal than 1")
 	}
-
+	log.Debugf("### TX Output: %s", txn.String())
 	return nil
 }
