+++ dpos/state/arbitrators.go
@@ -531,131 +531,6 @@ func (a *arbitrators) distributeDPOSReward(height uint32,
 	return
 }
 
-func (a *arbitrators) distributeWithNormalArbitratorsV2(height uint32, reward common.Fixed64) (
-	map[common.Uint168]common.Fixed64, common.Fixed64, error) {
-	if len(a.currentArbitrators) == 0 {
-		return nil, 0, errors.New("not found arbiters when " +
-			"distributeWithNormalArbitratorsV1")
-	}
-
-	roundReward := map[common.Uint168]common.Fixed64{}
-	totalBlockConfirmReward := float64(reward) * 0.25
-	totalTopProducersReward := float64(reward) - totalBlockConfirmReward
-	// Consider that there is no only CR consensus.
-	arbitersCount := len(a.chainParams.CRCArbiters) + a.chainParams.GeneralArbiters
-	individualBlockConfirmReward := common.Fixed64(
-		math.Floor(totalBlockConfirmReward / float64(arbitersCount)))
-	totalVotesInRound := a.CurrentReward.TotalVotesInRound
-	if len(a.chainParams.CRCArbiters) == len(a.currentArbitrators) {
-		// if no normal DPOS node, need to destroy reward.
-		roundReward[a.chainParams.DestroyELAAddress] = reward
-		return roundReward, reward, nil
-	}
-	rewardPerVote := totalTopProducersReward / float64(totalVotesInRound)
-
-	realDPOSReward := common.Fixed64(0)
-	for _, arbiter := range a.currentArbitrators {
-		ownerHash := arbiter.GetOwnerProgramHash()
-		rewardHash := ownerHash
-		var r common.Fixed64
-		if _, ok := a.crcArbiters[ownerHash]; ok {
-			r = individualBlockConfirmReward
-			m, ok := arbiter.(*crcArbiter)
-			if !ok || m.crMember.MemberState != state.MemberElected {
-				rewardHash = a.chainParams.DestroyELAAddress
-			} else {
-				pk := arbiter.GetOwnerPublicKey()
-				programHash, err := contract.PublicKeyToStandardProgramHash(pk)
-				if err != nil {
-					rewardHash = a.chainParams.DestroyELAAddress
-				} else {
-					rewardHash = *programHash
-				}
-			}
-		} else {
-			votes := a.CurrentReward.OwnerVotesInRound[ownerHash]
-			individualProducerReward := common.Fixed64(math.Floor(float64(
-				votes) * rewardPerVote))
-			r = individualBlockConfirmReward + individualProducerReward
-		}
-		roundReward[rewardHash] += r
-		realDPOSReward += r
-	}
-	for _, candidate := range a.currentCandidates {
-		ownerHash := candidate.GetOwnerProgramHash()
-		votes := a.CurrentReward.OwnerVotesInRound[ownerHash]
-		individualProducerReward := common.Fixed64(math.Floor(float64(
-			votes) * rewardPerVote))
-		roundReward[ownerHash] = individualProducerReward
-
-		realDPOSReward += individualProducerReward
-	}
-	// Abnormal CR`s reward need to be destroyed.
-	for i := len(a.currentArbitrators); i < arbitersCount; i++ {
-		roundReward[a.chainParams.DestroyELAAddress] += individualBlockConfirmReward
-	}
-	return roundReward, realDPOSReward, nil
-}
-
-func (a *arbitrators) distributeWithNormalArbitratorsV1(height uint32, reward common.Fixed64) (
-	map[common.Uint168]common.Fixed64, common.Fixed64, error) {
-	if len(a.currentArbitrators) == 0 {
-		return nil, 0, errors.New("not found arbiters when " +
-			"distributeWithNormalArbitratorsV1")
-	}
-
-	roundReward := map[common.Uint168]common.Fixed64{}
-	totalBlockConfirmReward := float64(reward) * 0.25
-	totalTopProducersReward := float64(reward) - totalBlockConfirmReward
-	individualBlockConfirmReward := common.Fixed64(
-		math.Floor(totalBlockConfirmReward / float64(len(a.currentArbitrators))))
-	totalVotesInRound := a.CurrentReward.TotalVotesInRound
-	if len(a.chainParams.CRCArbiters) == len(a.currentArbitrators) {
-		roundReward[a.chainParams.CRCAddress] = reward
-		return roundReward, reward, nil
-	}
-	rewardPerVote := totalTopProducersReward / float64(totalVotesInRound)
-
-	realDPOSReward := common.Fixed64(0)
-	for _, arbiter := range a.currentArbitrators {
-		ownerHash := arbiter.GetOwnerProgramHash()
-		rewardHash := ownerHash
-		var r common.Fixed64
-		if _, ok := a.crcArbiters[ownerHash]; ok {
-			r = individualBlockConfirmReward
-			m, ok := arbiter.(*crcArbiter)
-			if !ok || m.crMember.MemberState != state.MemberElected {
-				rewardHash = a.chainParams.DestroyELAAddress
-			} else {
-				pk := arbiter.GetOwnerPublicKey()
-				programHash, err := contract.PublicKeyToStandardProgramHash(pk)
-				if err != nil {
-					rewardHash = a.chainParams.DestroyELAAddress
-				} else {
-					rewardHash = *programHash
-				}
-			}
-		} else {
-			votes := a.CurrentReward.OwnerVotesInRound[ownerHash]
-			individualProducerReward := common.Fixed64(math.Floor(float64(
-				votes) * rewardPerVote))
-			r = individualBlockConfirmReward + individualProducerReward
-		}
-		roundReward[rewardHash] += r
-		realDPOSReward += r
-	}
-	for _, candidate := range a.currentCandidates {
-		ownerHash := candidate.GetOwnerProgramHash()
-		votes := a.CurrentReward.OwnerVotesInRound[ownerHash]
-		individualProducerReward := common.Fixed64(math.Floor(float64(
-			votes) * rewardPerVote))
-		roundReward[ownerHash] = individualProducerReward
-
-		realDPOSReward += individualProducerReward
-	}
-	return roundReward, realDPOSReward, nil
-}
-
 func (a *arbitrators) GetNeedConnectArbiters() []peer.PID {
 	a.mtx.Lock()
 	defer a.mtx.Unlock()
